requirements.txt
Package                       Version
----------------------------- ---------
asgiref                       3.4.1
backports.zoneinfo            0.2.1
certifi                       2021.10.8
cffi                          1.15.0
charset-normalizer            2.0.9
coreapi                       2.3.3
coreschema                    0.0.4
cryptography                  36.0.1
defusedxml                    0.7.1
diff-match-patch              20200713
Django                        3.2.10
django-appconf                1.0.5
django-imagekit               4.1.0
django-import-export          2.7.0
django-rest-knox              4.1.0
django-rest-passwordreset     1.2.1
djangorestframework           3.13.1
djangorestframework-simplejwt 5.0.0
drf-yasg                      1.20.0
et-xmlfile                    1.1.0
idna                          3.3
inflection                    0.5.1
itypes                        1.2.0
Jinja2                        3.0.3
MarkupPy                      1.14
MarkupSafe                    2.0.1
odfpy                         1.4.1
openpyxl                      3.0.9
packaging                     21.3
pilkit                        2.0
Pillow                        8.4.0
pip                           21.3.1
pycparser                     2.21
PyJWT                         2.3.0
pyparsing                     3.0.6
pytz                          2021.3
PyYAML                        6.0
requests                      2.26.0
ruamel.yaml                   0.17.18
ruamel.yaml.clib              0.2.6
setuptools                    47.1.0
six                           1.16.0
sqlparse                      0.4.2
tablib                        3.1.0
tzdata                        2021.5
uritemplate                   4.1.1
urllib3                       1.26.7
xlrd                          2.0.1
xlwt                          1.3.0
